클라 장비 아이콘 출력 오류 수정

+++ Source/Client/Assets/Scripts/UI/Popup/Game/UIInventoryPopup.cs
@@ -104,7 +104,7 @@ public class UIInventoryPopup : UIPopup
         for (GearType gearType = 0; gearType < GearType.END; ++gearType)
         {
             if(gear.Index[(int)gearType] > 0)
-                UpdateEquipSlotIcon(gearType, Managers.ItemData.GetGearIcon(Managers.Object.MyPlayer.Job, gearType, (byte)gear.Index[(int)gearType]), true);
+                UpdateEquipSlotIcon(gearType, Managers.ItemData.GetGearIcon(Managers.Object.MyPlayer.Job, gearType, gear.Index[(int)gearType]), true);
         }
     }
